Refactor to its own class, UiSynchronizer. Update description for better clarity. Add calls to its Initialize method.
